requirements.txt
Flask
requests
pymupdf
google-cloud-aiplatform